+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "dev": "vite build --watch",
     "build": "vite build",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "node test/index.js"
   },
   "repository": {
     "type": "git",
